Skeletal picture frame
This here is a basic picture frame but I cut out the middle of it and used longer than necessary pins to show off how the dowl roods in a picture framework. I also put in some pretty chamfers and even a few fillets to liven it up a bit.